LPS-196947 Inline AssetListEntryUsagesManager methods into ContentManager
